requirements.txt
appdirs==1.4.4
asgiref==3.5.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
distlib==0.3.1
Django==4.0.4
django-taggit==2.1.0
ffmpeg==1.4
filelock==3.0.12
fonttools==4.34.4
idna==2.10
importlib-metadata==4.11.4
kiwisolver==1.4.3
Markdown==3.3.6
matplotlib==3.5.2
mutagen==1.45.1
Naked==0.1.31
numpy==1.23.1
packaging==21.3
pandas==1.4.3
peewee==3.13.3
Pillow==9.1.0
pipenv==2020.11.15
psycopg2-binary==2.9.3
pycryptodome==3.15.0
Pygments==2.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==10.8.4
pytz==2021.1
PyYAML==6.0
rapidfuzz==1.4.1
requests==2.25.1
rich==10.3.0
sentry-sdk==0.19.4
shellescape==3.8.1
six==1.15.0
soupsieve==2.2.1
spotdl==3.6.1
spotipy==2.16.1
sqlparse==0.4.2
tqdm==4.61.1
tzdata==2022.1
Unidecode==1.2.0
urllib3==1.26.5
virtualenv==20.2.2
virtualenv-clone==0.5.4
youtube-dl==2021.6.6
ytmusicapi==0.17.3
zipp==3.8.0
